requirement.txt
# Requirements for Fire Detection API - Render.com
# Python 3.9+

Flask==2.3.3
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.0
gunicorn==21.2.0